Feature/optimize crud

+++ packages/express-app-users-api/app/services/users.services.js
@@ -1,12 +1,43 @@
 import { UserModel } from '../models';
+import { DefaultExternalUsersService } from '@josema-pereira/external-users';
+import { defaultConfig } from '../providers';
 
 export const UserServices = {
   getAll: async () => UserModel.find() || [],
+  getById: async (ids) => {
+    return new Promise(async (resolve, reject) => {
+      try {
+        const missingIds = [];
+        let users = await UserModel.find({ Id: { $in: ids } });
+        if (users.length !== ids.length) {
+          const {
+            users: { api },
+          } = defaultConfig;
+          const service = new DefaultExternalUsersService(api);
+          ids.forEach((id) => {
+            if (!users.find((user) => user.Id === id)) {
+              missingIds.push(id);
+            }
+          });
+          const missingUsers = await service.getUsers(missingIds);
+          users = users.concat(missingUsers);
+        }
+        resolve(users);
+      } catch (ex) {
+        reject({ message: ex });
+      }
+    });
+  },
   create: async (user) => {
     return new Promise(async (resolve, reject) => {
       try {
-        await UserModel.create({ ...user });
-        resolve({ message: 'User created' });
+        let existUser = await UserModel.findOne({ Id: user.Id });
+        if (existUser) {
+          resolve({ user: null, message: 'User not created. Duplicate Id' });
+        } else {
+          const newUser = await UserModel.create({ ...user });
+          resolve({ user: newUser });
+        }
       } catch (ex) {
         reject({ message: ex });
       }
@@ -15,10 +46,15 @@ export const UserServices = {
   update: async (id, newInfo) => {
     return new Promise(async (resolve, reject) => {
       try {
-        const user = await UserModel.findOneAndUpdate({ Id: id }, newInfo, {
-          new: true,
-        });
-        resolve({ message: 'User Update' });
+        let existUser = await UserModel.findOne({ Id: id });
+        if (existUser) {
+          const user = await UserModel.findOneAndUpdate({ Id: id }, newInfo, {
+            new: true,
+          });
+          resolve({ user });
+        } else {
+          resolve({ user: null, message: 'UserId not found' });
+        }
       } catch (ex) {
         reject({ message: ex });
       }
@@ -27,8 +63,13 @@ export const UserServices = {
   delete: async (id) => {
     return new Promise(async (resolve, reject) => {
       try {
-        const user = await UserModel.findOneAndDelete({ Id: id })
-        resolve({ message: 'User Deleted' });
+        let existUser = await UserModel.findOne({ Id: id });
+        if (existUser) {
+          const user = await UserModel.findOneAndDelete({ Id: id });
+          resolve({ user });
+        }else{
+          resolve({ user: null, message: 'UserId not found' });
+        }
       } catch (ex) {
         reject({ message: ex });
       }
